% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure_volume_data.R
\docType{data}
\name{pressure_volume_data}
\alias{pressure_volume_data}
\title{Pressure volume curve data}
\format{
A data frame with 160 rows and 8 variables
}
\usage{
pressure_volume_data
}
\description{
A dataset containing water potential and fresh mass measurements of repeatedly measured drying kohlrabi leaves subjected to
different soil moisture conditions during their growth (n = 6) and their saturated fresh mass and dry mass.
}
\details{
\itemize{
  \item date: Date of measurement
  \item treatment: Soil moisture conditions during the last 6 days of Kohlrabi growth (10-30 % (10 %) or 40-60 % (40 %))
  \item sample: Sample ID (1 - 12)
  \item fresh.mass.harvest: Fresh mass measured at harvest (12 h prior measurement of fresh.mass.saturated) (2.9813 - 7.1557)
  \item fresh.mass.saturated: Saturated fresh mass of the leaf in gramms (4.1276 - 7.0867)
  \item fresh.mass: Fresh mass of the leaf in gramms (2.7215 - 6.8246)
  \item dry.mass: Dry mass of the leaf in gramms (0.2937 - 0.7267)
  \item water.potential: Water potential of the leaf in MPa (-1.62 - -0.24)
}
}
\keyword{datasets}
