% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidityCheck.R
\name{ValidityCheck}
\alias{ValidityCheck}
\title{Ensures the validity of the input data}
\usage{
ValidityCheck(
  data,
  sample = FALSE,
  dry.mass = FALSE,
  fresh.mass.saturated = FALSE,
  fresh.mass = FALSE,
  water.potential = FALSE,
  RWD = FALSE
)
}
\arguments{
\item{data}{data frame containing the data to be checked}

\item{sample}{name of column containing the sample ID (default: sample)}

\item{dry.mass}{name of column containing the dry mass (g) (default: dry mass)}

\item{fresh.mass.saturated}{name of column containing the saturated fresh mass (g) (default: fresh.mass.saturated)}

\item{fresh.mass}{name of column containing the fresh mass (g) (default: fresh.mass)}

\item{water.potential}{name of column containing the water potential (MPa) (default: water.potential)}

\item{RWD}{name of column containing the relative water deficit (default: RWD)}
}
\value{
no return value
}
\description{
Ensures the validity of the input data
}
